Encode will quote strings containing colons or octothorpes
Otherwise Encode.toString produces YAML that fails to be decoded by
Decode.fromString.

It seems to me that better solution would be to improve decoder to allow
colons in strings as long as they are not immediately followed by space
or newline and hash symbols as long as they are not preceeded by a space
or newline. This is what https://yaml-multiline.info/ suggests. Quickly
reading YAML spec I'm not sure if that is legitimate though. In any case
I couldn't figure out how to make it. That's why I fell back to this
work-around.